# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateWorkspaceRequestDto": {
        "properties": {
          "locale": {
            "default": "en",
            "maxLength": 35,
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "type": "string"
          },
          "private": {
            "default": true,
            "type": "boolean"
          },
          "timezone": {
            "default": "UTC",
            "maxLength": 100,
            "type": "string"
          },
          "tone": {
            "maxLength": 40,
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "UpdateWorkspaceRequestDto": {
        "properties": {
          "name": {
            "maxLength": 160,
            "type": "string"
          },
          "private": {
            "type": "boolean"
          },
          "tone": {
            "maxLength": 40,
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "WorkspaceMemberResponseDto": {
        "properties": {
          "avatarUrl": {
            "nullable": true,
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "initials": {
            "type": "string"
          },
          "role": {
            "enum": [
              "OWNER",
              "MEMBER"
            ],
            "type": "string"
          },
          "userId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "displayName",
          "initials",
          "avatarUrl",
          "role"
        ],
        "type": "object"
      },
      "WorkspaceResponseDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "private": {
            "type": "boolean"
          },
          "role": {
            "enum": [
              "OWNER",
              "MEMBER"
            ],
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "tone": {
            "nullable": true,
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "tone",
          "private",
          "timezone",
          "locale",
          "role",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Versioned REST API for ClickFlow workspaces and projects.",
    "title": "ClickFlow API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspaces": {
      "get": {
        "operationId": "WorkspaceController_list",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      },
      "post": {
        "operationId": "WorkspaceController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}": {
      "delete": {
        "operationId": "WorkspaceController_archive",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      },
      "get": {
        "operationId": "WorkspaceController_get",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      },
      "patch": {
        "operationId": "WorkspaceController_update",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
...
    "/api/v1/workspaces/{workspaceId}/members": {
      "get": {
        "operationId": "WorkspaceController_members",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceMemberResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
...
  },
  "servers": [],
...
}